Rewrote hash code

+++ scracth/sha256.h
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 
 void init_hash(uint32_t h[8]);
-void sha256(const uint8_t data[], uint32_t length, uint32_t hash[8]);
+void sha256(const uint8_t data[], size_t length, uint32_t hash[8]);
 void sha256_next(const uint8_t data[], uint32_t length, uint32_t hash[8],
                  size_t total_length, uint8_t *tmpbuf);
 void sha256_stream(uint32_t hash[8], const uint8_t data[], uint64_t length);
